refactor: streamline branding type imports and improve organization
- Updated imports for `BrandingData` to reference the new location in `../types/branding`, enhancing code organization.
- Removed redundant type definitions from `brandingStore`, ensuring a single source of truth for `BrandingData`.
- Adjusted related hooks and tests to align with the new import structure, improving maintainability and clarity.
