+++ src/index.ts
@@ -257,7 +257,8 @@ export type { ThemeStore, ThemeState, ThemeActions } from './store/themeStore';
 
 // Branding Store
 export { useBrandingStore } from './store/brandingStore';
-export type { BrandingData, BrandingState } from './store/brandingStore';
+export type { BrandingState } from './store/brandingStore';
+export type { BrandingData } from './types/branding';
 
 // Storage Keys
 export { KEYS, FEATURE_FLAGS_KEYS } from './utils/keys';
